lib/src: DC link Tap/limit setting to avoid inrush fault
DC link current channel TAP/limit setting to 128/1.0
avoiding system faults due to inrush current. As the
the system ramps on speed there a inrush current for
shot interval. This requires a box car filter in the system
to avoid false tripping. A similar feature will be implemented
in software. This commit is a fallback in place to SW implementation.