lib: Add event monitoring through EventManager class
This adds EventManager class to handle the event of the system.
Incomming faults and interrupts are handled through event monitoring
thread. The thread will wait on epoll events for the events and call
the callbacks / update the event's shadow status as registered.
It will get the necessary descriptor from the respective driver class of the event.

Signed-off-by: Pallav Joshi <pallav.joshi@amd.com>