lib: event handling: Exit thread gracefully.
The epoll_wait is intentionally called without any timeout
to get maximum thread yield. But with this approach, it will
be impossible to exit the thread during the destruction or application
exit; as epoll_wait will be block indefinitely waiting for events on
registered FDs.

To allow graceful thread exit, an additional fd is added. This is the
eventfd, which can be triggered for event/interrupt by the software (
from destructor in this case). This signal is to be issued after setting
mMonitorRunning=false. The thread will come out of the epoll_wait loop
block and will not do the next iteration as Running is false.

Signed-off-by: Pallav Joshi <pallav.joshi@amd.com>