lib: foc: Use seperate enum for the operation modes
As mentioned in previous todo, we need to seperate out operation mode
for the motor control library and acutal operation mode being set in
the FOC driver. The operation mode for the library = hw mode setting +
gate drive setting.

All HW mode setting may not be exposed to the user (e.g. FixedAngle is
internal mode used just for the calibration).

MotorControlImpl translates the user operation modes to the foc
operation modes.

Signed-off-by: Pallav Joshi <pallav.joshi@amd.com>