CMake: Add option to skip the post installation
On target the post installation will take care of
creating soft link for the config file in the /etc/motor-control
directory.

When installed through the debian or other packaging system, the post
installation should be handled by packaging system and not the cmake.

So add a flag to conditionally skip the post installation.

Signed-off-by: Pallav Joshi <pallav.joshi@amd.com>